feat(treemap): move stories to the dedicated storybook workspace
